Migrated remaining routes
