+++ backend/src/routes/api/library.js
@@ -1,0 +1,16 @@
+// Template file for api route
+var router = require('express').Router();
+var getDb = require('../../pg-db').getDb;
+var db = getDb();
+
+router.route('/').get((req, res) => {
+    db.any('select * from cloudbreak_cuisine.external_bundles')
+        .then(data => {
+            res.json(data);
+        })
+        .catch(error => {
+            console.log('ERROR:', error)
+        })
+});
+
+module.exports = router;
